pyproject.toml
[tool.poetry]
name = "nuvla-job-engine"
version = "3.9.5"
description = "Nuvla Job Engine."
license = "Apache-2.0"
authors = ["SixSq SA <support@sixsq.com>"]
maintainers = ["Ignacio Penas <nacho@sixsq.com>"]
readme = "README.md"
repository = "https://github.com/nuvla/job-engine"
classifiers = [
        "Development Status :: 5 - Production/Stable",
        "License :: OSI Approved :: Apache Software License",
        "Intended Audience :: Developers",
        "Intended Audience :: System Administrators",
        "Operating System :: OS Independent",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.11",
        "Topic :: Software Development"]
packages = [{ include = "job_engine" }, { include = "scripts"}]


[tool.poetry.dependencies]
python = "^3.11"
nuvla-api = "^3.0.9"
docker = "^7.0.0"
pyyaml = "^6.0.1"
statsd = "^4.0.1"
packaging = "^23.2"
importlib-metadata = "^7.0.1"


[tool.poetry.group.server.dependencies]
elasticsearch = "^7"
boto3 = "^1.34.40"
kazoo = "^2.6.0"
requests = "2.27.1"


[tool.poetry.group.test.dependencies]
mock = "^5.1.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
coverage = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
